job-tracker/requirements.txt
flask
python-dotenv
psycopg2-binary
requests
langgraph
langchain
scikit-learn